Refactor world_entry_methods into hybrid 2-level hierarchy
Split the flat 24-file world_entry_methods/ into a semantic hierarchy:
- player_load/ (2 files: core, meta) — player data loading
- inventory/ (3 files: core, grant, move_) — inventory operations
- vendor/ (14 files organized by operation) — vendor store system

Benefits:
- Semantic grouping signals code organization to new developers
- Maintains max ~300 lines per file; average now ~200 lines
- Shorter import paths: vendor::purchase:: vs vendor_purchase::
- Optimized for LLM context: grouped files loaded together
- Full backward compatibility via mod.rs re-exports

Also refactored cell_methods/player from monolithic 772-line file
into 9 focused modules (combat, interaction, vendor, world, crafting, social, etc).

Schema updates:
- Added 3 new player columns (bandolier_slot, training_points, applied_science_points)
- Added 4 inventory test items for character 62
- New vendor_store_smoke.sql test suite for full vendor stack

Rust infrastructure:
- Added 5 new BaseToCellMsg variants for bandolier/inventory sync
- Added vendor_entity field to CellEntity struct
- Added ON_STORE_OPEN/ON_STORE_UPDATE method indices
- Updated entity method dispatch to handle new message types
