+++ build.log
@@ -1,0 +1,27 @@
+   Compiling proc-macro2 v1.0.106
+   Compiling quote v1.0.44
+   Compiling serde v1.0.228
+   Compiling serde_core v1.0.228
+   Compiling typenum v1.19.0
+   Compiling generic-array v0.14.7
+   Compiling windows_x86_64_msvc v0.53.1
+   Compiling parking_lot_core v0.9.12
+error: linker `C:\Users\Steve\.rustup\toolchains\stable-x86_64-pc-windows-msvc\lib\rustlib\x86_64-pc-windows-msvc\bin\rust-lld.exe` not found
+  |
+  = note: The system cannot find the path specified. (os error 3)
+
+note: the msvc targets depend on the msvc linker but `link.exe` was not found
+
+note: please ensure that Visual Studio 2017 or later, or Build Tools for Visual Studio were installed with the Visual C++ option
+
+note: VS Code is a different product, and is not sufficient
+
+error: could not compile `serde` (build script) due to 1 previous error
+warning: build failed, waiting for other jobs to finish...
+error: could not compile `serde_core` (build script) due to 1 previous error
+error: could not compile `proc-macro2` (build script) due to 1 previous error
+error: could not compile `quote` (build script) due to 1 previous error
+error: could not compile `generic-array` (build script) due to 1 previous error
+error: could not compile `typenum` (build script) due to 1 previous error
+error: could not compile `windows_x86_64_msvc` (build script) due to 1 previous error
+error: could not compile `parking_lot_core` (build script) due to 1 previous error
