+++ crates/services/src/base/world_entry_appearance.rs
@@ -90,7 +90,7 @@ pub(crate) async fn handle_on_client_ready(
     );
 
     // Query saved missions from DB before sending InitPlayerState
-    let saved_missions = super::world_entry_player::query_saved_missions(
+    let saved_missions = super::world_entry_methods::query_saved_missions(
         db_pool, pending.player_id,
     ).await;
 
@@ -116,6 +116,8 @@ pub(crate) async fn handle_on_client_ready(
             world_name: pending.world_name.clone(),
             saved_missions,
             abilities,
+            active_bandolier_slot: 0,
+            bandolier_items: vec![],
         }).await;
     }
