+++ crates/services/src/base/world_entry_methods/inventory/mod.rs
@@ -1,0 +1,7 @@
+pub mod core;
+pub mod grant;
+pub mod move_;
+
+pub use core::{send_full_inventory_update, handle_remove_inventory_item};
+pub use grant::{normalize_item_ids, item_allows_container, handle_grant_item};
+pub use move_::handle_move_inventory_item;
