+++ crates/services/src/cell/cell_methods/inventory.rs
@@ -13,6 +13,23 @@ pub const REPAIR_ITEM_REQUEST: u16 = 40;
 pub const REQUEST_ACTIVE_SLOT_CHANGE: u16 = 41;
 pub const REQUEST_AMMO_CHANGE: u16 = 42;
 
+/// Fire a weapon attack event for an equipped item.
+/// Called after a ranged weapon attack is launched.
+pub async fn fire_equipped_weapon_attack_event(
+    entity_id: u32,
+    target_id: i32,
+    is_ranged: bool,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    tracing::debug!(
+        entity_id,
+        target_id,
+        is_ranged,
+        "fireEquippedWeaponAttackEvent"
+    );
+}
+
 pub async fn dispatch(
     entity_id: u32,
     method_index: u16,
