+++ crates/services/src/cell/cell_methods/player/mod.rs
@@ -1,0 +1,11 @@
+pub mod constants;
+mod dispatch;
+pub mod combat;
+pub mod interaction;
+pub mod vendor;
+pub mod world;
+pub mod crafting;
+pub mod social;
+
+pub use dispatch::dispatch;
+pub use constants::*;
